AI/quiz/requirements.txt
flask>=2.0.0
huggingface-hub
python-dotenv
flask-cors
